Moved App contents to new Dashboard component to clean up code. Implemented logout feature to delete bearer token on sign out and display the login page again.
