+++ src/components/Dashboard/Dashboard.css
@@ -1,0 +1,40 @@
+.wrapper {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.header {
+  position: fixed;
+  top: 0;
+  width: 100vw;
+  margin-top: 0vh;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.body {
+  position: fixed;
+  top: 20vh;
+  width: 70vw;
+}
+.taskbar-button {
+  width: 16vw;
+  height: 4vh;
+  padding: 0px;
+  margin: 1vh;
+  background-color: white;
+  color: #282c34;
+  text-align: center;
+  text-decoration: none;
+  font-size: calc(6px + 2vmin);
+  border: none;
+  border-radius: 20px;
+}
